revert types
